Refactor test code
